project-2/requirements.txt
langchain-google-genai
langchain
langchain_experimental
langchain_openai
python-dotenv
numpy
pandas
matplotlib
tabulate